// packages/transition/src/enums/transition-enums.ts
export enum TransitionChannels {
    main = 'transition:main',
    ended = 'transition:ended',
    progress = 'transition:progress',
}

export enum TransitionModes {
    enter = 'enter',
    leave = 'leave',
}

export enum TransitionStatus {
    ENTERING = 0,
    ENTERED = 1,
    LEAVING = 2,
    LEFT = 3,
}